Include partitioned cookies in export
Firefox 141 added support for partitioned cookies[^1] by default. In
order for extensions to retrieve these cookies, they need to pass a
`partitionKey` object, which has been supported since Firefox 94[^2].

While we could consider adding options to filter which cookies are
exported, for now the easy win is to select all partitions, so this
commit just passes an empty object to select them all.

There's no support for partitions in the cookies.txt format or in any of
the tools that consume it, but it can be useful to see the partition for
debugging, so this commit includes it in a comment.

[^1]: https://developer.mozilla.org/en-US/docs/Web/Privacy/Guides/Privacy_sandbox/Partitioned_cookies
[^2]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/cookies/getAll#browser_compatibility